+++ openrouteservice/src/main/java/org/heigit/ors/routing/configuration/RouteProfileConfiguration.java
@@ -33,6 +33,7 @@ public class RouteProfileConfiguration {
 	private Double maximumDistanceDynamicWeights = 0.0;
 	private Double maximumDistanceAvoidAreas = 0.0;
 	private Double maximumDistanceAlternativeRoutes = 0.0;
+	private Double maximumDistanceRoundTripRoutes = 0.0;
 	private Integer maximumWayPoints = 0;
 	private boolean instructions = true;
 	private boolean optimize = false;
@@ -210,6 +211,14 @@ public class RouteProfileConfiguration {
 
 	public void setMaximumDistanceAlternativeRoutes(Double maximumDistanceAlternativeRoutes) {
 		this.maximumDistanceAlternativeRoutes = maximumDistanceAlternativeRoutes;
+	}
+
+	public Double getMaximumDistanceRoundTripRoutes() {
+		return maximumDistanceRoundTripRoutes;
+	}
+
+	public void setMaximumDistanceRoundTripRoutes(Double maximumDistanceRoundTripRoutes) {
+		this.maximumDistanceRoundTripRoutes = maximumDistanceRoundTripRoutes;
 	}
 
 	public void setMaximumWayPoints(Integer value)